Include tzdata as implicit runtime dependency.
tidyr universe package `lubridate` depends on information in `/usr/share/zoneinfo` that is provided by the tzdata library added through apk.

This occurred in the R 3.* world (Greenline), but I wanted to propose introducing this into R 4.* containers while we're thinking about it.  And as something to do during a cloudformation deploy!